Updated Meta Utils.
