+++ trade/utils/__init__.py
@@ -4,5 +4,4 @@ from trade.utils.gsheet_util import GSheetCredType
 from trade.utils.gsheet_util import GSpreadManager as GoogleSheet
 from trade.utils.log_configurator import LogConfig as Logger
 from trade.utils.log_configurator import LoggingType
-from trade.utils.meta_utils import SingletonMeta, Utilities
 from trade.utils.network_tools import DownloadTools
