--- trade/utils/meta_utils/__init__.py
@@ -1,2 +1,0 @@
-from trade.utils.meta_utils.master_meta import UtilityEnabler as Utilities
-from trade.utils.meta_utils.singleton_meta import SingletonMeta
