--- trade/utils/meta_utils/singleton_meta.py
@@ -1,8 +1,0 @@
-class SingletonMeta(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            cls._instances[cls] = super().__call__(*args, **kwargs)
-
-        return cls._instances[cls]
