+++ trade/utils/telegram_api.py
@@ -2,12 +2,20 @@ import asyncio
 from datetime import datetime
 from functools import cache
 from os import getenv
-
 import telegram
 
-from trade.utils.meta_utils.singleton_meta import SingletonMeta
 
 TELEGRAM_SIGNATURE = "\n Generated on {0}."
+
+
+class SingletonMeta(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super().__call__(*args, **kwargs)
+
+        return cls._instances[cls]
 
 
 class TelegramBot(metaclass=SingletonMeta):
